lexbridge-ai-main/brain_module/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain_module"
version = "0.1.0"
description = "Multi-source Q&A reasoning and synthesis layer over data_module"
requires-python = ">=3.10"
dependencies = [
    # retrieval / reranking
    "sentence-transformers>=2.6.1",
    "torch>=2.2.2",
    # LightRAG client
    "httpx>=0.27.0",
    "aiohttp>=3.9.5",
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.1.0",
    # LLM clients
    "openai>=1.35.0",
    "litellm>=1.40.0",
    # schema / utilities
    "langdetect>=1.0.9",
    "pydantic>=2.7.0",
    "tenacity>=8.3.0",
    "python-dotenv>=1.0.0",
    # caching
    "redis>=5.0.4",
    # evaluation
    "ragas>=0.1.14",
    "datasets>=2.19.0",
    "lightrag-hku[api]>=1.4.14",
    "pyjwt>=2.12.1",
    "bcrypt>=5.0.0",
    "ollama>=0.6.1",
    "neo4j>=6.1.0",
    "data-module",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27.0"]
# Whisper + pydub for POST /audio/transcribe (ffmpeg must be on PATH)
# edge-tts for POST /audio/speak (streaming multilingual TTS)
audio = ["openai-whisper>=20231117", "pydub>=0.25.1", "edge-tts>=7.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["brain_module*"]

[tool.uv.sources]
data-module = { path = "../data_module", editable = true }
